Group controller fixes

+++ MewaAppBackend.WebApi/Handlers/Group/DeleteGroupCommandHandler.cs
@@ -4,7 +4,7 @@ using MewaAppBackend.WebApi.Commands.Group;
 
 namespace MewaAppBackend.WebApi.Handlers.Group
 {
-    public class DeleteGroupCommandHandler : IRequestHandler<DeleteGroupCommand, Unit>
+    public class DeleteGroupCommandHandler : IRequestHandler<DeleteGroupCommand, DeleteGroupCommandResult>
     {
         private readonly IUnitOfWork unitOfWork;
         private readonly IRepository<Model.Model.Group> groupRepository;
@@ -14,12 +14,17 @@ namespace MewaAppBackend.WebApi.Handlers.Group
             groupRepository = unitOfWork.Repository<Model.Model.Group>();
         }
 
-        public async Task<Unit> Handle(DeleteGroupCommand request, CancellationToken cancellationToken)
+        public async Task<DeleteGroupCommandResult> Handle(DeleteGroupCommand request, CancellationToken cancellationToken)
         {
             var group = groupRepository.GetDetail(g => g.Id == request.Id);
+            if (group == null)
+            {
+                return new DeleteGroupCommandResult { Message = $"Grupa o id {request.Id} nie istnieje!", Success = false };
+            }
             groupRepository.Delete(group);
             unitOfWork.SaveChanges();
-            return Unit.Value;
+
+            return new DeleteGroupCommandResult { Success = true };
         }
     }
 }
